Scrolls and looker
Added in lightning scroll consumale that targets nearest enemy in 20 blocks and
the ability to look round the map using the slash key. This has been implemented
using the mouse to move the cursor around and the keyboard movement controls. Shift,
ctrl, alt all modify speed of keyboard navigation in this mode.!
!